+++ pcs/common/reports/messages.py
@@ -7415,3 +7415,25 @@ class AddRemoveCannotSpecifyAdjacentItemWithoutItemsToAdd(ReportItemMessage):
             f"Cannot specify adjacent {item} '{self.adjacent_item_id}' without "
             f"{items} to add"
         )
+
+
+@dataclass(frozen=True)
+class ResourceStonithCommandsMismatch(ReportItemMessage):
+    """
+    Command designed for a resource has been used with a stonith resource/agent
+    or the other way around.
+
+    not_accepted_type -- description of an entity not being accepted anymore
+    command_to_use_instead -- identifier of a command to use instead
+    """
+
+    not_accepted_type: str
+    command_to_use_instead: Optional[types.PcsCommand] = None
+    _code = codes.RESOURCE_STONITH_COMMANDS_MISMATCH
+
+    @property
+    def message(self) -> str:
+        return (
+            f"Ability of this command to accept {self.not_accepted_type} is "
+            "deprecated and will be removed in a future release."
+        )