+++ pcs/lib/cib/fencing_topology.py
@@ -23,7 +23,7 @@ from pcs.common.reports import (
 from pcs.common.reports import codes as report_codes
 from pcs.common.reports import has_errors
 from pcs.common.reports.item import ReportItem
-from pcs.lib.cib.stonith import is_stonith_resource
+from pcs.lib.cib.resource.stonith import is_stonith_resource
 from pcs.lib.cib.tools import find_unique_id
 from pcs.lib.errors import LibraryError
 from pcs.lib.pacemaker.state import _Element as StateElement